requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langgraph>=0.1.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.0.5
alpaca-trade-api>=3.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
google-generativeai>=0.5.0
aiofiles>=23.0.0
websockets>=10.0,<11.0
